update expense tracker

+++ src/main/java/com/example/InstaPay_Travel_Tours/service/ExpenseService.java
@@ -9,6 +9,7 @@ import java.util.List;
 
 @Service
 public class ExpenseService {
+
     @Autowired
     private ExpenseRepository expenseRepository;
 
@@ -16,15 +17,15 @@ public class ExpenseService {
         return expenseRepository.findAll();
     }
 
-    public void saveExpense(Expense expense) {
-        expenseRepository.save(expense);
+    public Expense saveExpense(Expense expense) {
+        return expenseRepository.save(expense);
     }
 
-    public Expense getExpenseById(Long id) {
-        return expenseRepository.findById(id).orElse(null);
-    }
-
-    public void deleteExpenseById(Long id) {
-        expenseRepository.deleteById(id);
+    public boolean deleteExpense(Long id) {
+        if (expenseRepository.existsById(id)) {
+            expenseRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
